+++ src/config/local.example.js
@@ -1,0 +1,3 @@
+export default {
+    SOUNDCLOUD_CLIENT_ID: "YOUR SOUNDCLOUD API KEY"    
+  };
