+++ src/constants/constants.js
@@ -1,0 +1,3 @@
+import config from "../config/local";
+
+export const SOUNDCLOUD_CLIENT_ID = config.SOUNDCLOUD_CLIENT_ID;
